@@ -1,4 +1,4 @@
 import * as pine from '@pinefile/pine';
 import { register } from './experiments';
 
-module.exports = register(pine);
+export default register(pine);
